# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TaskFlow API
  description: |
    API สำหรับระบบจัดการงานและทรัพยากรองค์กร (Task & Resource Management System)
    
    ## Authentication
    API ใช้ JWT Bearer Token สำหรับการยืนยันตัวตน
    ```
    Authorization: Bearer <token>
    ```
    
    ## Rate Limiting
    - Login endpoint: 5 requests/minute per IP
    - Other endpoints: ไม่จำกัด (ในเวอร์ชันปัจจุบัน)
    
    ## Roles
    - **ADMIN** - ผู้ดูแลระบบ
    - **CUSTOMER_SERVICE** - ฝ่ายบริการลูกค้า
    - **FINANCE_LEADER** - หัวหน้าฝ่ายการเงิน
    - **FINANCE** - เจ้าหน้าที่การเงิน
    - **SALES_LEADER** - หัวหน้าฝ่ายขาย
    - **SALES** - พนักงานขาย
    - **HEAD_TECH** - หัวหน้าแผนกช่าง
    - **LEADER** - หัวหน้าทีม
    - **TECH** - ช่าง
    
  version: 1.0.0
  contact:
    name: TaskFlow Support
    email: support@taskflow.com
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://taskflow.example.com/api
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token ที่ได้จาก /auth/login
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        message:
          type: string
        error:
          type: string
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
...
    Role:
      type: string
      enum:
        - ADMIN
        - CUSTOMER_SERVICE
        - FINANCE_LEADER
        - FINANCE
        - SALES_LEADER
        - SALES
        - HEAD_TECH
        - LEADER
        - TECH
...
    TaskStatus:
      type: string
      enum:
        - WAITING
        - IN_PROGRESS
        - DONE
        - CANCELLED
...
    CreateTaskRequest:
      type: object
      required:
        - title
        - startDate
        - endDate
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        location:
          type: string
          maxLength: 500
        customerName:
          type: string
          maxLength: 200
        customerPhone:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        startTime:
          type: string
          default: '09:00'
        endTime:
          type: string
          default: '17:00'
        subUnitId:
          type: string
        priority:
          type: integer
          minimum: 1
          maximum: 5
          default: 1
...
  responses:
    Unauthorized:
      description: ไม่ได้ยืนยันตัวตน
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error: กรุณาเข้าสู่ระบบ
    Forbidden:
      description: ไม่มีสิทธิ์เข้าถึง
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error: ไม่มีสิทธิ์เข้าถึง
    BadRequest:
      description: ข้อมูลไม่ถูกต้อง
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error: กรุณากรอกข้อมูลให้ครบถ้วน
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: ดึงรายชื่อผู้ใช้
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 100
        - name: search
          in: query
          schema:
            type: string
        - name: role
          in: query
          description: comma-separated roles
          schema:
            type: string
        - name: departmentId
          in: query
          schema:
            type: string
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: รายชื่อผู้ใช้
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PaginatedResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Users]
      summary: สร้างผู้ใช้ใหม่
      description: ต้องมีสิทธิ์ canManageUsers
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [employeeId, email, password, name, role]
              properties:
                employeeId:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
                name:
                  type: string
                phone:
                  type: string
                role:
                  $ref: '#/components/schemas/Role'
                departmentId:
                  type: string
                subUnitId:
                  type: string
      responses:
        '200':
          description: สร้างผู้ใช้สำเร็จ
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ==================== Tasks ====================
  /tasks:
    get:
      tags: [Tasks]
      summary: ดึงรายการงาน
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TaskStatus'
        - name: subUnitId
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: รายการงาน
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Tasks]
      summary: สร้างงานใหม่
      description: ต้องมีสิทธิ์ canCreateTasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '200':
          description: สร้างงานสำเร็จ
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...